+++ database/migrations/2021_05_27_082335_create_navigation_trees_table.php
@@ -13,15 +13,17 @@ class CreateNavigationTreesTable extends Migration
      */
     public function up()
     {
-        Schema::create('trees', function (Blueprint $table) {
-            $table->id();
-            $table->string('handle');
-            $table->string('type');
-            $table->string('initialPath')->nullable();
-            $table->string('locale')->nullable();
-            $table->json('tree')->nullable();
-            $table->timestamps();
-        });
+        if (config('statamic-eloquent-driver.trees.enabled', true)) {
+            Schema::create('trees', function (Blueprint $table) {
+                $table->id();
+                $table->string('handle');
+                $table->string('type');
+                $table->string('initialPath')->nullable();
+                $table->string('locale')->nullable();
+                $table->json('tree')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -31,6 +33,8 @@ class CreateNavigationTreesTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('trees');
+        if (config('statamic-eloquent-driver.trees.enabled', true)) {
+            Schema::dropIfExists('trees');
+        }
     }
 }
